Set up OAuth 2.0 process for Spotify
The green button on the index page will now start the authentication
process with Spotify. It is done using the Authorization Code Flow using
NextJS API routes.

Currently, the end of the index page will just leave the user back at
'/#' with the necessary information in the GET query string.

Token refreshing is not yet implemented.