Allow state connection in / to be undefined
This is to represent the initial state when the browser has not
established whether it is connected to Spotify (e.g. because the call to
Spotify's API server has not responded yet).

The intention is for such a state to be representable using e.g.
spinner / progress bar so that users will not be confused.