Add some comments to spotify.ts; add establish static method
Begin adding comments to SpotifyConnection as this class will probably
become very complicated in the future.

Add an 'establish' static method, which is used to establish and verify
a connection to a user's Spotify account given the access token and
refresh tokens. Also automatically uses the refresh token as necessary.
I will eventually make the constructor private to enforce that all
created instances of SpotifyConnection to have been verified.