Refactor authentication to use SpotifyConnection#establish
The React#useEffect hook in index.tsx will now delegate the process of
instantiating the SpotifyConnection instance to a new function,
verifyAuthTokens, in utils/auth.ts.

This function is a rename of the old handleAuthSuccess function (as it
also handles failure cases) and makes use of the new
SpotifyConnection#establish static method to establish a connection with
the given authentication tokens.

The constructor of SpotifyConnection is now private.