SpotifyConnection will try refreshing token for all network tasks
Create a new method SpotifyConnection#tryRefresh, which handles the task
of refreshing the access token, and returns a Promise of the new access token.
The _accessToken field is also automatically updated. The
onAccessTokenChange callback supplied by SpotifyConnection#establish is
now a private field and will also be called with the new access token.

This is called by SpotifyConnection#get whenever there is an HTTP 401
error. Hence, all network tasks using get will attempt to refresh the
access token on HTTP 401.

This deprecates SpotifyConnection#verify as it is redundant with
SpotifyConnection#getUserProfile (which has the same capabilites).
Hence, it is removed and SpotifyConnection#establish will now use
SpotifyConnection#getUserProfile instead.