Implement search function for ImportWizard
Users can now search for playlists they wish to import.

This necessitated converting the 'playlist' object in the store from an
array to a map (with the key being the playlist URI) as search will not
preserve the index of each StagedPlaylist object, which is necessary for
several 'playlist' operations.

Renamed 'pagedPlaylists' in some components to 'playlistContainer'.

Moved PlaylistCard to its own module.

Removed the checkbox concerning duplicate tracks as handling individual
tracks is no longer in the scope of ImportWizard.

Create a (for now, decorative) button to Import playlists selected in
the dialog.

Create a button to load the next page of playlists.