Use chainable thunks instead of callbacks
Define a new type ChainableThunk<T> =
AppThunk<Promise<PayloadAction<T>>>

The thunk creators loadNextPage and loadAll have been revised to make
use of chainable thunks instead of callbacks (to avoid callback hell).

(I originally wanted to do this, but I thought it was impossible. It
turns out the solution is to cast the return value of useDispatch() to
AppDispatch.)