Begin implementing SongManager features
Create the slice which contains the state for the SongManager.

The Import button on ImportWizard now has an action - importing
playlists.

SongManager doesn't render the imported playlists just yet - it only
shows the number of imported playlists.

Re-add the checkbox controlling whether to remove duplicate tracks, but
under SongManager (instead of ImportWizard).

Add a selector to choose how to render the playlists. Currently, it only
has one mode - by playlist.