Tokens are now passed through cookies instead of querystring
During authentication, the tokens are passed from the server to the
client through cookies instead of the querystring. The client will parse
these cookies accordingly.

This is so that browser history (back/forward/etc.) works as expected.

Only the refresh token is permanently stored; the access token will be
removed immediately on page load.

Also added a informational alert about using cookies.